+++ src/main/java/Heuristic/PopulationBasedMetaHeuristic/SetOfMethods/IteratedLocalSearch/DBHC.java
@@ -16,13 +16,12 @@ public abstract class DBHC extends HillClimb {
         for(int i=0;i<problem.getDepthOfSearch();i++){
             int[] indices= IntStream.range(0,problem.getNumberOfVariables()).toArray();
             int[] perm=shuffle(indices,random);
-            int currentCost=this.problem.getObjectiveFunctionValue(index);
-            int originCost=currentCost;
+            double currentCost=this.problem.getObjectiveFunctionValue(index);
+            double originCost=currentCost;
             for (int j = 0; j < perm.length; j++) {
                 this.problem.bitFlip(index,perm[j]);
-                int candidateCost = deltaEvaluation(index);
+                double candidateCost = deltaEvaluation(index);
                 if (acceptMove(currentCost, candidateCost)) {
-
                     currentCost = candidateCost;
 
                 } else {
